remove superflous test file
